requirements-dev.txt
# Инструменты разработки для проекта portfolio-system-architect
# Установка: pip install -r requirements-dev.txt

# Листинг и форматирование
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0

# Проверка типов
mypy>=1.0.0

# Безопасность
bandit>=1.7.0
safety>=2.0.0

# Документация
pydocstyle>=6.3.0
docformatter>=1.7.0

# Тестирование (дополнительные инструменты)
pytest-xdist>=3.0.0
pytest-html>=3.2.0
pytest-mock>=3.10.0
coverage>=7.0.0
pytest-playwright
pytest-cov
requests

# Зависимости для конкретных компонентов
streamlit>=1.32.0
fastapi[all]>=0.68.0

# Инструменты для работы с данными
pandas[performance]>=1.3.2
openpyxl>=3.0.0

# Дополнительные утилиты
pre-commit>=3.0.0
commitizen>=2.0.0
